se agrega la validacion para evitar el error al intentar borrar el directio temp en caso de no existir

+++ PDADesktop/Classes/Utils/Fileutils.cs
@@ -115,7 +115,10 @@ namespace PDADesktop.Classes.Utils
             string tempFilePathExtended = TextUtils.ExpandEnviromentVariable(tempFilePath);
             try
             {
-                Directory.Delete(@tempFilePathExtended, true);
+                if (Directory.Exists(@tempFilePathExtended))
+                {
+                    Directory.Delete(@tempFilePathExtended, true);
+                }
             }
             catch (IOException ex)
             {
